WIP #242 predict disease on offline mode
