+++ disease_detector_app/android/app/build.gradle
@@ -49,10 +49,15 @@ android {
         applicationId "com.example.disease_detector_app"
         // You can update the following values to match your application needs.
         // For more information, see: https://docs.flutter.dev/deployment/android#reviewing-the-gradle-build-configuration.
-        minSdkVersion 21
+        minSdkVersion 26
         targetSdkVersion flutter.targetSdkVersion
         versionCode flutterVersionCode.toInteger()
         versionName flutterVersionName
+    }
+
+    lintOptions {
+        warningsAsErrors true
+        disable 'Deprecation'
     }
 
     buildTypes {
@@ -61,6 +66,11 @@ android {
             // Signing with the debug keys for now, so `flutter run --release` works.
             signingConfig signingConfigs.debug
         }
+    }  
+    
+    aaptOptions {
+        noCompress "tflite"
+        noCompress "lite"
     }
 }
