+++ disease_detector_app/lib/screens/account_page/account_screen_offline.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+class AccountScreenOffline extends StatelessWidget {
+  const AccountScreenOffline({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(
+      child: Text(
+        'No Internet Connection',
+      ),
+    );
+  }
+}
