+++ disease_detector_app/lib/screens/history/history_screen_offline.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+class HistoryScreenOffline extends StatelessWidget {
+  const HistoryScreenOffline({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(
+      child: Text(
+        'No Internet Connection',
+      ),
+    );
+  }
+}
