+++ disease_detector_app/linux/flutter/generated_plugin_registrant.cc
@@ -8,6 +8,7 @@
 
 #include <file_selector_linux/file_selector_plugin.h>
 #include <flutter_localization/flutter_localization_plugin.h>
+#include <tflite_flutter_plus/tflite_flutter_plus_plugin.h>
 #include <url_launcher_linux/url_launcher_plugin.h>
 
 void fl_register_plugins(FlPluginRegistry* registry) {
@@ -17,6 +18,9 @@ void fl_register_plugins(FlPluginRegistry* registry) {
   g_autoptr(FlPluginRegistrar) flutter_localization_registrar =
       fl_plugin_registry_get_registrar_for_plugin(registry, "FlutterLocalizationPlugin");
   flutter_localization_plugin_register_with_registrar(flutter_localization_registrar);
+  g_autoptr(FlPluginRegistrar) tflite_flutter_plus_registrar =
+      fl_plugin_registry_get_registrar_for_plugin(registry, "TfliteFlutterPlusPlugin");
+  tflite_flutter_plus_plugin_register_with_registrar(tflite_flutter_plus_registrar);
   g_autoptr(FlPluginRegistrar) url_launcher_linux_registrar =
       fl_plugin_registry_get_registrar_for_plugin(registry, "UrlLauncherPlugin");
   url_launcher_plugin_register_with_registrar(url_launcher_linux_registrar);
