+++ disease_detector_app/package/service/LICENSE
@@ -1,0 +1,1 @@
+TODO: Add your license here.
