+++ disease_detector_app/package/service/lib/service.dart
@@ -1,0 +1,8 @@
+
+export 'model_service/model_service.dart';
+
+/// A Calculator.
+class Calculator {
+  /// Returns [value] plus 1.
+  int addOne(int value) => value + 1;
+}
