+++ disease_detector_app/pubspec.yaml
@@ -12,11 +12,13 @@ dependencies:
   flutter:
     sdk: flutter
   cupertino_icons: ^1.0.8
+  service: 
+    path: package/service
   iconsax: ^0.0.8
   intl: ^0.19.0
   provider: ^6.1.2
   url_launcher: ^6.2.6
-  flutter_native_splash: ^2.4.0
+  flutter_native_splash: ^2.2.16
   flutter_screenutil: ^5.9.1
   fluttertoast: ^8.2.5
   firebase_core: ^2.31.1
@@ -36,15 +38,19 @@ dependencies:
   path_provider: ^2.1.3
   flutter_pdfview: ^1.3.2
   flutter_localization: ^0.2.1
+  tflite_flutter: ^0.10.4
+  tflite_flutter_helper_plus: ^0.0.1
   flutter_localizations:
     sdk: flutter
+  google_fonts: ^6.2.1
+  connectivity_plus: ^6.0.5
 
 dev_dependencies:
   flutter_lints: ^4.0.0
   flutter_test:
     sdk: flutter
 
-flutter:
+flutter:  
   uses-material-design: true
   generate: true
 
@@ -54,6 +60,7 @@ flutter:
     - assets/logo/
     - assets/images/onboarding/
     - assets/icons/
+    - assets/model/
 
   fonts:
     - family: Poppins
