+++ disease_detector_app/windows/flutter/generated_plugins.cmake
@@ -4,15 +4,18 @@
 
 list(APPEND FLUTTER_PLUGIN_LIST
   cloud_firestore
+  connectivity_plus
   file_selector_windows
   firebase_auth
   firebase_core
   firebase_storage
   flutter_localization
+  tflite_flutter_plus
   url_launcher_windows
 )
 
 list(APPEND FLUTTER_FFI_PLUGIN_LIST
+  tflite_flutter
 )
 
 set(PLUGIN_BUNDLED_LIBRARIES)
